+++ src/modules/core/GameConfig.h
@@ -85,6 +85,7 @@ constexpr const char *VoxformatColorAsFloat = "voxformat_colorasfloat";
 constexpr const char *VoxformatWithtexcoords = "voxformat_withtexcoords";
 constexpr const char *VoxformatPointCloudSize = "voxformat_pointcloudsize";
 constexpr const char *VoxformatTransform = "voxformat_transform_mesh";
+constexpr const char *VoxformatOptimize = "voxformat_optimize";
 constexpr const char *VoxformatFillHollow = "voxformat_fillhollow";
 constexpr const char *VoxformatVoxelizeMode = "voxformat_voxelizemode";
 constexpr const char *VoxformatVXLNormalType = "voxformat_vxlnormaltype";