+++ src/modules/voxel/SurfaceExtractor.cpp
@@ -13,40 +13,40 @@ namespace voxel {
 
 SurfaceExtractionContext buildCubicContext(const RawVolume *volume, const Region &region, ChunkMesh &mesh,
 										   const glm::ivec3 &translate, bool mergeQuads, bool reuseVertices,
-										   bool ambientOcclusion) {
+										   bool ambientOcclusion, bool optimize) {
 	return SurfaceExtractionContext(volume, getPalette(), region, mesh, translate, SurfaceExtractionType::Cubic,
-									mergeQuads, reuseVertices, ambientOcclusion);
+									mergeQuads, reuseVertices, ambientOcclusion, optimize);
 }
 
 SurfaceExtractionContext buildMarchingCubesContext(const RawVolume *volume, const Region &region, ChunkMesh &mesh,
-												   const palette::Palette &palette) {
+												   const palette::Palette &palette, bool optimize) {
 	return SurfaceExtractionContext(volume, palette, region, mesh, glm::ivec3(0), SurfaceExtractionType::MarchingCubes,
-									false, false, false);
+									false, false, false, optimize);
 }
 
 void extractSurface(SurfaceExtractionContext &ctx) {
 	if (ctx.type == SurfaceExtractionType::MarchingCubes) {
 		voxel::Region extractRegion = ctx.region;
 		extractRegion.shrink(-1);
-		voxel::extractMarchingCubesMesh(ctx.volume, ctx.palette, extractRegion, &ctx.mesh);
+		voxel::extractMarchingCubesMesh(ctx.volume, ctx.palette, extractRegion, &ctx.mesh, ctx.optimize);
 	} else {
 		voxel::Region extractRegion = ctx.region;
 		if (ctx.volume->region() == extractRegion) {
 			extractRegion.shiftUpperCorner(1, 1, 1);
 		}
 		voxel::extractCubicMesh(ctx.volume, extractRegion, &ctx.mesh, ctx.translate, ctx.mergeQuads, ctx.reuseVertices,
-								ctx.ambientOcclusion);
+								ctx.ambientOcclusion, ctx.optimize);
 	}
 }
 
 voxel::SurfaceExtractionContext createContext(voxel::SurfaceExtractionType type, const voxel::RawVolume *volume,
 											  const voxel::Region &region, const palette::Palette &palette,
 											  voxel::ChunkMesh &mesh, const glm::ivec3 &translate, bool mergeQuads,
-											  bool reuseVertices, bool ambientOcclusion) {
+											  bool reuseVertices, bool ambientOcclusion, bool optimize) {
 	if (type == voxel::SurfaceExtractionType::MarchingCubes) {
-		return voxel::buildMarchingCubesContext(volume, region, mesh, palette);
+		return voxel::buildMarchingCubesContext(volume, region, mesh, palette, optimize);
 	}
-	return voxel::buildCubicContext(volume, region, mesh, translate, mergeQuads, reuseVertices, ambientOcclusion);
+	return voxel::buildCubicContext(volume, region, mesh, translate, mergeQuads, reuseVertices, ambientOcclusion, optimize);
 }
 
 } // namespace voxel