+++ src/modules/voxel/SurfaceExtractor.h
@@ -20,9 +20,9 @@ enum class SurfaceExtractionType { Cubic, MarchingCubes, Max };
 struct SurfaceExtractionContext {
 	SurfaceExtractionContext(const RawVolume *_volume, const palette::Palette &_palette, const Region &_region,
 							 ChunkMesh &_mesh, const glm::ivec3 &_translate, SurfaceExtractionType _type,
-							 bool _mergeQuads, bool _reuseVertices, bool _ambientOcclusion)
+							 bool _mergeQuads, bool _reuseVertices, bool _ambientOcclusion, bool _optimize)
 		: volume(_volume), palette(_palette), region(_region), mesh(_mesh), translate(_translate), type(_type),
-		  mergeQuads(_mergeQuads), reuseVertices(_reuseVertices), ambientOcclusion(_ambientOcclusion) {
+		  mergeQuads(_mergeQuads), reuseVertices(_reuseVertices), ambientOcclusion(_ambientOcclusion), optimize(_optimize) {
 	}
 	const RawVolume *volume;
 	const palette::Palette &palette; // used only for MarchingCubes
@@ -33,13 +33,14 @@ struct SurfaceExtractionContext {
 	const bool mergeQuads;		 // used only for Cubic
 	const bool reuseVertices;	 // used only for Cubic
 	const bool ambientOcclusion; // used only for Cubic
+	const bool optimize;
 };
 
 SurfaceExtractionContext buildCubicContext(const RawVolume *volume, const Region &region, ChunkMesh &mesh,
 										   const glm::ivec3 &translate = glm::ivec3(0), bool mergeQuads = true,
-										   bool reuseVertices = true, bool ambientOcclusion = true);
+										   bool reuseVertices = true, bool ambientOcclusion = true, bool optimize = false);
 SurfaceExtractionContext buildMarchingCubesContext(const RawVolume *volume, const Region &region, ChunkMesh &mesh,
-												   const palette::Palette &palette);
+												   const palette::Palette &palette, bool optimize = false);
 
 void extractSurface(SurfaceExtractionContext &ctx);
 
@@ -47,6 +48,6 @@ voxel::SurfaceExtractionContext createContext(voxel::SurfaceExtractionType type,
 											  const voxel::Region &region, const palette::Palette &palette,
 											  voxel::ChunkMesh &mesh, const glm::ivec3 &translate,
 											  bool mergeQuads = true, bool reuseVertices = true,
-											  bool ambientOcclusion = true);
+											  bool ambientOcclusion = true, bool optimize = false);
 
 } // namespace voxel