+++ src/modules/voxel/private/MarchingCubesSurfaceExtractor.h
@@ -16,6 +16,6 @@ struct ChunkMesh;
 
 // Also known as: "3D Contouring", "Marching Cubes", "Surface Reconstruction"
 void extractMarchingCubesMesh(const RawVolume *volume, const palette::Palette &palette, const Region &region,
-							  ChunkMesh *result);
+							  ChunkMesh *result, bool optimize);
 
 } // namespace voxel