+++ docs/CHANGELOG.md
@@ -12,6 +12,8 @@ Known [issues](https://github.com/vengi-voxel/vengi/issues?q=is%3Aissue+is%3Aope
 
 General:
 
+   - Add meshoptimizer support (new cvar `voxformat_optimize`)
+
 ## 0.0.31 (2024-05-09)
 
 General: